src/github.com/dnvriend/deck/deck_test.go
package deck

import (
	"testing"
	"os"
)

const TestFileName = "_decktesting"

func TestNewDeck(t *testing.T) {
	d := NewDeck()
	if len(d) != 16 {
		t.Errorf("Expected deck length of 16, but got %v", len(d))
	}

	if d[0] != "Ace of Spades" {
		t.Errorf("Expected first card of Ace of Spades, but got %v", d[0])
	}

	if d[len(d)-1] != "Four of Clubs" {
		t.Errorf("Expected first card of Four of Clubs, but got %v", d[len(d)-1])
	}
}

func TestSaveToDeckAndNewDeckFromFile(t *testing.T) {
	os.Remove(TestFileName)
	deck := NewDeck()
	deck.SaveToFile(TestFileName)
	loadedDeck := NewDeckFromFile(TestFileName)
	if len(loadedDeck) != 16 {
		t.Errorf("Expected 16 cards in deck, got %v", len(loadedDeck))
	}
	os.Remove(TestFileName)
}

